+++ src/common.js
@@ -1,3 +1,5 @@
+import chalk from 'chalk'
+
 // Needed to use ES5 inheritance, because of issues with Error subclassing for Babel
 export function RunJSError (message) {
   this.name = 'RunJSError'
@@ -5,3 +7,21 @@ export function RunJSError (message) {
 }
 RunJSError.prototype = Object.create(Error.prototype)
 RunJSError.prototype.constructor = RunJSError
+
+export const logger = {
+  debug: (...args) => {
+    console.log(chalk.blue(...args))
+  },
+  info: (...args) => {
+    console.log(chalk.bold(...args))
+  },
+  log: (...args) => {
+    console.log(...args)
+  },
+  warning: (...args) => {
+    console.warn(chalk.yellow(...args))
+  },
+  error: (...args) => {
+    console.error(chalk.red(...args))
+  }
+}
